requirements.txt
aioredis==2.0.1
alembic==1.15.2
asyncpg==0.30.0
fastapi==0.115.12
greenlet==3.2.2
motor==3.7.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pymongo==4.13.0
python-dotenv==1.1.0
SQLAlchemy==2.0.41
uvicorn==0.34.2
